Updated Directions
Based on a standard address, 2000 Pennington Road, and set restaurant locations (Randomly selected from google maps). Provides the map and a set of directions in the sidebar next to it. Spans the entire area where the map is placed. Can be placed anywhere in the body by putting all the code in the body tags in this doc somewhere in the body of our web pages. The head code needs to be put anywhere in the head part of the html for the website, specifically for pages we plan to use google maps.